jcf-io.c (read_zip_member): Cast to long in comparison with signed value.
	* jcf-io.c (read_zip_member): Cast to long in comparison with
	signed value.

	* jvspec.c (lang_specific_driver): Initialize variables.

	* mangle.c (find_compression_record_match): Likewise.

	* typeck.c (build_null_signature): Provide static prototype.  Mark
	parameter with ATTRIBUTE_UNUSED.

	* verify.c (verify_jvm_instructions): Initialize variable.

From-SVN: r41671

+++ gcc/java/jcf-io.c
@@ -192,7 +192,7 @@ DEFUN(read_zip_member, (jcf, zipd, zipf),
 	      jcf->read_ptr = jcf->buffer;
 	      jcf->read_end = jcf->buffer_end;
 	      if (lseek (zipf->fd, zipd->filestart, 0) < 0
-		  || read (zipf->fd, jcf->buffer, zipd->size) != zipd->size)
+		  || read (zipf->fd, jcf->buffer, zipd->size) != (long) zipd->size)
 	        return -2;
 	    }
 	  else
@@ -213,7 +213,7 @@ DEFUN(read_zip_member, (jcf, zipd, zipf),
 	      d_stream.next_in = buffer;
 	      d_stream.avail_in = zipd->size;
 	      if (lseek (zipf->fd, zipd->filestart, 0) < 0
-		  || read (zipf->fd, buffer, zipd->size) != zipd->size)
+		  || read (zipf->fd, buffer, zipd->size) != (long) zipd->size)
 		return -2;
 	      /* Handle NO_HEADER using undocumented zlib feature.
                  This is a very common hack.  */
